feat(v2): split v2 services blueprint
relates #7, #8, #12
